Fixed bug where cur_time was always after the snapshot_to time
The issue was the mocked network server and the game were using the same cur_time, but in reality, the server will be ahead of the game. So, the server and the game were on the same time when the server should've been ahead. To fix this, the mocked network server uses cur_server_time from platformer::time_t to have its own time stream. cur_indepdent_time was also removed and other time stuff since they are useless in this scenario.