Changed finish state of smoothing function to only use time ratio
This is because smoothing at the same cur_time can be applied for multiple objects, so if one object is close to target, it doesn't mean another object at same time is as well. Time ratio is better way of ensuring all objects finishing smoothing at same time during extrapolation.