Added pipes to the game
Added a pipe struct to represent a pipe specifying its bottom position. Read it in from the level map file and increased collision detection extent to be 4 grids rather than 2 to ensure pipes are always tested with the player since their position is specified based on bottom of pipe rather than center.